trabalho/src/requirements.txt
kafka-python==1.4.7
psycopg2-binary==2.9.9
boto3==1.28.49
numpy<2
pandas==2.2.2
pyarrow==12.0.1